이길주 - UI-TOGGLE GROUP 까지 구현

+++ 공통/MapleStory_v2.3/Renderer2D.h
@@ -29,7 +29,7 @@ public:
 	void Render(Matrix3x2F _mat, ID2D1RenderTarget *_pRT);
 
 	void AddBitmap(ID2D1Bitmap* _bitmap);
-	void ChangeBitmap(int _index);
+	void ChangeBitmap(size_t _index);
 
 	inline void ClearBitmap()
 	{
@@ -44,6 +44,7 @@ public:
 	inline void SetImgRT(D2D1_RECT_F _rt)
 	{
 		m_ImgRTList[m_BitmapIndex] = _rt;
+
 	}
 	inline D2D1_RECT_F GetImgRT()
 	{
